requirements.txt
django
djangorestframework
django_extensions
scipy
sympy
celery
psycopg2